refactor(blueprint): split build.rs into build/ and csv_loader.rs into table + typing
Phase 1 of the blueprint-input-formats program, behaviour-preserving. The
2148-line build.rs sat 352 lines under the god-file ceiling and the next
phase adds an input registry and per-format dispatch, so it is cut along
its own phase banners first: build/{mod,specs,cache,manual,nodes,fk,
junction,table_ops}.rs. csv_loader.rs splits into the string table
(RawCsv, chunk reader, misparse tally) and the typing half (declared arms,
inference, list/date cells) so the coming producer seam has one file to
replace. Widened visibility is pub(super) only; lib.rs's
blueprint::build re-export path is unchanged. 102 blueprint tests pass
under their new module paths.
